assignment

+++ EAD-ShoppingCart/src/com/metacube/shoppingcart/Model/Product.java
@@ -5,9 +5,9 @@ public class Product {
 	int product_code;
 	String product_type;
 	String product_name;
-	float product_price;
+	double product_price;
 	
-	public Product(int product_code, String product_type, String product_name, float product_price){
+	public Product(int product_code, String product_type, String product_name, double product_price){
 		this.product_name = product_name;
 		this.product_code = product_code;
 		this.product_price = product_price;
@@ -26,7 +26,7 @@ public class Product {
 		return product_name;
 	}
 
-	public float getProduct_price() {
+	public double getProduct_price() {
 		return product_price;
 	}
